commit: [fe] refactor : `ContentWrapper` 를 `components/`로 이동 - `UserMatchingPage` 에서도 재사용하기 위해 옮깁니다!
