+++ client/src/components/common/ContentWrapper.tsx
